ARM64: defconfig: mi8996: increase HZ value to 300
QCOM set the default value to 100, but that value is good for a server, not an interactive device

From Google docs:
Increase the CONFIG_HZ value.
Historically, the value has been set to 100 on arm and arm64 platforms.
However, this is an accident of history and is not a good value to use for interactive devices.
CONFIG_HZ=100 means that a jiffy is 10ms long, which means that load balancing between CPUs may take 20ms (two jiffies) to happen.
This can significantly contribute to jank on a loaded system.
Recent devices (Nexus 5X, Nexus 6P, Pixel, and Pixel XL) have shipped with CONFIG_HZ=300.
This should have a negligible power cost while significantly improving runnable times.
If you do see significant increases in power consumption or performance issues after changing CONFIG_HZ, it's likely one of your drivers is using a timer based on raw jiffies instead of milliseconds and converting to jiffies.
This is usually an easy fix (see the patch that fixed kgsl timer issues on Nexus 5X and 6P when converting to CONFIG_HZ=300).
Finally, we've experimented with CONFIG_HZ=1000 on Nexus/Pixel and found it offers a noticeable performance and power reduction due to decreased RCU overhead.
With those two changes alone, a device should look much better for UI thread runnable time under load.

Change-Id: I27aa4610dcc249ce877a1c1e15a4e768cdfe6df4
Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>